13 add detail song when played

+++ main.c
@@ -6,30 +6,7 @@
 #include "sorting.h"
 
 #define MAX_LEN 128
-void SpotifyText(){
-    printf("\033[1;32m"); // Set text color to bright green
-    printf("\n          ██████████                                                                            ████████████████████████████   ████████████   ");
-    printf("\n        ██████████████████                                                                      ████████████████████████████   ████████████   ");
-    printf("\n     ████████████████████████                                                                   ████████████████████████████   ████████████   ");
-    printf("\n    ██████████████████████████                                                                  ████████████████████████████   ████████████   ");
-    printf("\n  ██████████████████████████████                                                                                ████████████                  ");
-    printf("\n █████                  █████████        █████████                            ████  ███         ████████████    ████████████                  ");
-    printf("\n █████      █████           █████       ███    ███                                  ███         ████████████    ████████████                  ");
-    printf("\n█████████████████████████    █████      ███         ███ ██████     ████████    ███ ████████     ████████████    ████████████                  ");
-    printf("\n███████               ████████████      ████████    ████    ████  ████   ████  ███  ████        ████████████    ████████████                  ");
-    printf("\n███████   ████████        ████████         ███████  ███      ███ ███      ████ ███  ███         ████████████    ████████████                  ");
-    printf("\n████████████████████████   ███████              ███ ███      ███ ███      ████ ███  ███         ████████████    ████████████                  ");
-    printf("\n ██████               ███████████       ███     ███ ████    ████  ███    ████  ███  ████        ████████████    ████████████                  ");
-    printf("\n ████████████████████    ████████       ██████████  ███████████    █████████   ███   █████      ████████████    ████████████                  ");
-    printf("\n  ██████████████████████████████                    ███                                         ████████████    ████████████                  ");
-    printf("\n    ██████████████████████████                      ███                                         ████████████    ████████████                  ");
-    printf("\n     ████████████████████████                                                                   ████████████    ████████████                  ");
-    printf("\n        ██████████████████                                                                      ████████████    ████████████                  ");
-    printf("\n           ████████████                                                                         ████████████    ████████████                  ");
-    printf("\n\n\n\n");
-    printf("\033[0m"); // Reset text color to default
 
-}
 
 void printSongInsidePlaylist(Playlist* playlist, int index) {
     Playlist* curr = findPlaylistByIndex(playlist, index);
@@ -38,24 +15,24 @@ void printSongInsidePlaylist(Playlist* playlist, int index) {
         int len = strlen(curr->playlistName);
         int width = len > 159 ? len : 159;
         int textLength = strlen(curr->playlistName);
-        int leftPadding = (width - textLength) / 2; // Calculate left padding
-        int rightPadding = width - textLength - leftPadding; // Calculate right padding
+        int leftPadding = (width - textLength) / 2;
+        int rightPadding = width - textLength - leftPadding - 52;
         if(song != NULL){
-            printf("+%.*s+\n", 200, "=========================================================================================================================================================================");
+            printf("+%.*s+\n", 148, "=====================================================================================================================");
             printf("|\033[1;32m");
             printf("%*sPlaylist: %s%*s\033[0m|\n", leftPadding, "", curr->playlistName, rightPadding, "");
             // printf("");
-            printf("+%.*s+\n", 200, "=========================================================================================================================================================================");
+            printf("+%.*s+\n", 148, "=====================================================================================================================");
 
             // Header tabel
-            printf("| %-50s | %-50s | %-50s | %-8s | %-20s |\n", "Title", "Singer", "Album", "Duration", "Status");
-            printf("+----------------------------------------------------+----------------------------------------------------+----------------------------------------------------+----------+\n");
+            printf("| %-25s | %-25s | %-25s | %-8s | %-20s |\n", "Title", "Singer", "Album", "Duration", "Status");
+            printf("+---------------------------+---------------------------+---------------------------+----------+----------------------\n");
 
             // Baris data tabel
             while (song != NULL) {
-                printf("| %-50s | %-50s | %-50s | %-8.2f | %-20s |\n", song->title, song->singer, song->album, song->time, song->status);
+                printf("| %-25s | %-25s | %-25s | %-8.2f | %-20s |\n", song->title, song->singer, song->album, song->time, song->status);
                 song = song->next;
-                printf("+%.*s+\n", 200, "=========================================================================================================================================================================");
+                printf("+%.*s+\n", 148, "=====================================================================================================================");
             }
         }else{
             printf("\n\033[0;37;41mNo songs in this playlist.\033[0m\n");
@@ -82,13 +59,12 @@ void printPlaylist(Playlist* playlist) {
 }
 
 Playlist* playlistMenu(Playlist* playlist) {
-    char choice, extra;
-    char title[50], singer[50], album[50], remove[50], play[50], url[256], status[20], command[256], path[50];
+    char choice;
+    char title[50], singer[50], album[50], remove[50], play[50], url[256], status[20], command[256], path[64];
     float duration;
     int idx, idxtoremove, idxtoplay;
 
     do {
-        // SpotifyText();
         printf("\033[1;32m");
         printf("\033[1m\n-- PLAYLIST MANAGER --\033[0m\n");
         printf("\033[0m");
@@ -99,7 +75,6 @@ Playlist* playlistMenu(Playlist* playlist) {
         printf("e. Back to main menu\n");
         printf("Enter your choice: ");
         scanf(" %c", &choice);
-        // getchar();
 
         if (getchar() != '\n') {
             system("clear");
@@ -153,7 +128,6 @@ Playlist* playlistMenu(Playlist* playlist) {
                             printf("\n\033[0;37;41mThe singer is required.\033[0m\n");
                             break;
                         }
-
                         printf("Enter the album of the song to add: ");
                         fgets(album, sizeof(album), stdin); 
                         album[strcspn(album, "\n")] = '\0';
@@ -161,14 +135,6 @@ Playlist* playlistMenu(Playlist* playlist) {
                             strcpy(album, "Unknown");
                             album[strcspn(album, "\n")] = '\0';
                         }
-
-                        // printf("Enter the duration of the song to add with format (0.00): ");
-                        // if (scanf("%f%c", &duration, &extra) == 2 && extra == '\n');
-                        // else {
-                        //     printf("\n\033[0;37;41mInvalid input. must be a float.\033[0m\n");
-                        //     while ((extra = getchar()) != '\n' && extra != EOF);
-                        //     break;
-                        // }
 
                         printf("Enter the youtube url of the song to add \n(If filled, the music will be available to play): ");
                         fgets(url, sizeof(url), stdin); 
@@ -178,12 +144,13 @@ Playlist* playlistMenu(Playlist* playlist) {
                             strcpy(url, "-");
                             url[strcspn(url, "\n")] = '\0';
                         }else{                         
-                            printf("processing download song from youtube url...");   
-                            char* songName = escape(title);
+                            printf("processing download song from youtube url...\n");   
+                            // char* songName = escape(title);
+                            snprintf(path, sizeof(path), "songs/%s.mp3",title );
+                            char* songName = escape(path);
                             char* link = escape(url);
-                            snprintf(command, sizeof(command), "yt-dlp -q -x --audio-format mp3 --audio-quality 0 -o './songs/%s.mp3' %s", songName,link);
+                            snprintf(command, sizeof(command), "yt-dlp -q -x --audio-format mp3 --audio-quality 0 -o %s %s", songName,link);
                             system(command);
-                            snprintf(path, sizeof(path),"./songs/%s.mp3",title);
                             duration = getSongDuration(path);
                             if(duration >0.00){
                                 strcpy(status, "Available to play");
@@ -192,14 +159,11 @@ Playlist* playlistMenu(Playlist* playlist) {
                                 strcpy(status, "Unavailable to play");
                             }
                         }
-
                         status[strcspn(status, "\n")] = '\0';
                         strip(title);
                         strip(singer);
                         strip(album);
                         strip(url);
-                        system("clear");
-                        SpotifyText();
                         addSongToPlaylist(playlist, idx, title, singer, album, duration, url, status);
                         sortByTitle(playlist);
                     }else{
@@ -210,15 +174,11 @@ Playlist* playlistMenu(Playlist* playlist) {
                         printf("\n\033[0;37;41mInvalid number.\033[0m\n");    
                     }            
                     break;
-                
-                
                 case 'c':
                     system("clear");
                     SpotifyText();
                     printPlaylist(playlist);
                     printf("enter the number of the playlist that contains the song you want to delete: ");
-                    // scanf(" %d",&idxtoremove);
-                    // getchar();
                     if(scanf(" %d",&idxtoremove)){
                         getchar();
                         Playlist* curr=findPlaylistByIndex(playlist,idxtoremove);
@@ -226,7 +186,6 @@ Playlist* playlistMenu(Playlist* playlist) {
                             printf("\n\033[0;37;41mInvalid number, playlist not exists.\033[0m\n");
                             break;
                         }
-                        // printf("%s",curr->song->title);
                         if(curr->song!=NULL){
                             printSongInsidePlaylist(playlist,idxtoremove);
                             printf("Enter the name of the song to remove: ");
@@ -251,11 +210,9 @@ Playlist* playlistMenu(Playlist* playlist) {
                     break;
                 case 'd':
                     system("clear");
-                    SpotifyText();   
+                    SpotifyText();
                     printPlaylist(playlist);
                     printf("enter the number of playlist that contains the song you want to play: ");
-                    // scanf(" %d", &idxtoplay);
-                    // getchar();
                     if(scanf("%d",&idxtoplay)){
                         getchar();
                         Playlist* curr=findPlaylistByIndex(playlist,idxtoplay);
@@ -264,14 +221,13 @@ Playlist* playlistMenu(Playlist* playlist) {
                             break;
                         }
 
-                        if(curr->song!=NULL){ 
+                        if(curr->song!=NULL){
+
                             printSongInsidePlaylist(playlist, idxtoplay);
                             printf("enter the name of the song you want to play: ");
                             fgets(play, sizeof(play), stdin); 
                             play[strcspn(play, "\n")] = '\0';
                             strip(play);
-                            system("clear");
-                            SpotifyText();
                             playSong(playlist, idxtoplay, play);
                         }else{
                             printf("\n\033[0;37;41mPlaylist is empty.\033[0m\n"); 
@@ -284,7 +240,6 @@ Playlist* playlistMenu(Playlist* playlist) {
                         }
                         printf("\n\033[0;37;41mInvalid number.\033[0m\n");   
                     }
-
                     break;
                 case 'e':                           
                     system("clear");
@@ -296,6 +251,7 @@ Playlist* playlistMenu(Playlist* playlist) {
                     SpotifyText();
                     printPlaylist(playlist);
                     printf("\n\033[0;37;41mInvalid choice. Please try again.\033[0m\n");
+                    break;
             }
         }
     } while (choice != 'e');
@@ -308,7 +264,6 @@ Playlist* dashboardMenu(Playlist* playlist){
     char playlistName[50];
 
     do {
-        // SpotifyText();
         printf("\033[1;32m");
         printf("\033[1m\n-- SONG APP MENU --\033[0m\n");
         printf("\033[0m");
@@ -335,6 +290,10 @@ Playlist* dashboardMenu(Playlist* playlist){
                     printf("Enter playlist name: ");
                     fgets(playlistName, sizeof(playlistName), stdin); 
                     playlistName[strcspn(playlistName, "\n")] = '\0';
+                    if(playlistName[0] == '\0'){
+                            printf("\n\033[0;37;41mPlaylist name is required.\033[0m\n");
+                            break;
+                    }
                     strip(playlistName);
                     playlist = addNewPlaylist(playlist, playlistName);                
                     break;
@@ -361,7 +320,13 @@ Playlist* dashboardMenu(Playlist* playlist){
                         printf("enter the number of the playlist you want to delete: ");                        
                         if(scanf(" %d",&idxtoremove)){
                             getchar();
-                            playlist = deletePlaylist(playlist, idxtoremove);
+                            if(findPlaylistByIndex(playlist,idxtoremove)){
+                                playlist = deletePlaylist(playlist, idxtoremove);                                
+                            }else{
+                                printf("\n\033[0;37;41mInvalid number, playlist not exists.\033[0m\n");
+                                break;
+                            }
+
                         }else{
                             char ch;
                             while ((ch = getchar()) != '\n' && ch != EOF) {
@@ -372,9 +337,7 @@ Playlist* dashboardMenu(Playlist* playlist){
                         
                     }
                     break;
-
-                case 'd':
-                       
+                case 'd':                       
                     system("clear");
                     SpotifyText();
                     if(playlist == NULL){
@@ -384,7 +347,13 @@ Playlist* dashboardMenu(Playlist* playlist){
                         printf("Which playlist you want to save? (choose number of playlist): ");
                         if(scanf(" %d", &save)){
                             getchar();
-                            savePlaylist(playlist, save);
+                            if(findPlaylistByIndex(playlist,save)){
+                                savePlaylist(playlist, save);
+                            }else{
+                                printf("\n\033[0;37;41mInvalid number, playlist not exists.\033[0m\n");
+                                break;
+                            }
+
                         }else{
                             char ch;
                             while ((ch = getchar()) != '\n' && ch != EOF) {
@@ -402,10 +371,6 @@ Playlist* dashboardMenu(Playlist* playlist){
                         printf("Enter playlist name: ");
                         fgets(playlistName, sizeof(playlistName), stdin); 
                         playlistName[strcspn(playlistName, "\n")] = '\0';
-                        
-                    
-                        // SpotifyText();
-                        strip(playlistName);
                         playlist = readPlaylist(playlist,playlistName);
                     }
                     break;
@@ -425,9 +390,6 @@ Playlist* dashboardMenu(Playlist* playlist){
 }
 
 int main(){
-
-    // float duration = getSongDuration("blue.mp3");
-    // printf("%f",duration);
     Playlist* playlists = NULL;
     SpotifyText();
     playlists = dashboardMenu(playlists);
